refactor(store): lift the shared insight-audit helpers to the crate root
parse_error_payload, existing_keys and RAW_PAYLOAD_MAX_CHARS carry nothing
character-specific. All three insight chains need them, and the human chain
importing from character_insight would misstate where they belong.
